json: enable JSON schema versioning
In a future commit I'm going to introduce a new JSON schema
("version 2"), so set the ground-work for that here.  There are
capabilities that are unable to be expressed in the current format, such
as prioritization of two groups both have Blink settings but at
different rates, but we want to retain backwards compatibility to any
existing configs.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: Ib4cb5761de1f80567a775cb15e8da921f539037e