Reset lamps after reboot
In case lamp test has been trigerred and reboot takes place,
even after reboot the lamps are found to be in the on state.

The commit switches off the lamps in case such situation is
found and restores the LEDs to the state in which they were
before lamp test was trigerred.

The change was done against "PE00DKX2".

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>