Reset lamps after reboot
In case lamp test has been trigerred and reboot takes place,
even after reboot the lamps are found to be in the on state.

The commit switches off the lamps in case such situation is
found and restores the LEDs to the state in which they were
before lamp test was trigerred.

The change was done against "PE00DKX2".

Test:
- Take the bmc to standby state.
- Power LED on panel will be in blinking state.
- Turn on lamp test from panel.
- Kill the phosphor-ledmanager service.
- Service restarts, finds all the led in on state.
- All the leds are swiched off and power button return to
blinking state.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>