Set operational status for guarded DIMMs
The script checks for guarded DIMMs and Processors and sets the
operational status as false for those FRUs.
This change was required to re-establish the LED state for
guarded DIMMs and PROCs in ase it got cleared by clear_all_fault
script in the boot process.

It does not set the status in case cores are guarded.
The script does not operate in chassis on condition.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>