Select config json based on device tree
The commit implements change to read model from the /sys path and
pick config json based on that.

The change was done so that phosphor-ledmanager can get the config
file using these data without having to wait for any other servcie.

The change therefore reduces the probablity of having a failure
where phosphor-logging needs to trigger LED at a very initial
stage of the boot and LED manager is not yet up, waiting for config
file.

The change only executes when LED_USE_JSON is defined.

The commit though preserves the existing flow to handle the case
where model may be missing in config json or the entry is not
intended in the config json.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>