Persistence the led group asserted value by default
Currently, only IBM's P10 machine will persist the Asserted attribute
value of ledGroup by default. But for other systems, most do not need
to perform persistence actions.

The intent behind this commit is to add a switch to control the
persistence function, and it is enabled by default in CI and disabled
by default in Yocto.

Tested: Regardless of whether `persistent-led-asserted` is enabled or
disabled, the ledmanager process works fine.

Signed-off-by: Jinu Joy Thomas <jinu.joy.thomas@in.ibm.com>