led tool: toggle fault LEDs
The commit implements base framework for led tool with an option
to toggle fault LEDs based on the functional status of the FRU.

The tool will be auto triggered with different optiosn on BMC
boot and system power on, driven via system-d service files.

Tool can be used via command line as well.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>