Set LEDs state for guarded FRUs
The commit implements an option to set LEDs state for FRUs which
are marked as guarded.

This is required as all fault LEDs associated with FRUs gets cleared
on a BMC reboot at chassis off state. To avoid losing LEDs which were
lit up due to guard records, this will be excuted via service file.

Signed-off-by: sunny srivastava <sunny@sunnys-MacBook-Pro-2.local>