Call PIM to update the functional property (#83)
During power-on, clear-all-fault-leds service clears all the fault
leds by setting the functional property of the corresponding led FRU
to true. While setting the property, it calls the busctl set-property
method, which updates the data only on DBus and data is not persisted.

When BMC is rebooted while host is running, on system come back, the
stale data of the FRU which has functional as false will get updated on
the DBus, which marks that FRU as critical on ASMI.

Changes made to use Phosphor Inventory Manager(PIM) notify method to
update the functional property, which publish the data on Dbus and
persist the data.