Notify PIM to update functional property (#84)
Functional property of a FRU which is populated under the Phosphor
Inventory Manager(PIM) service will have persisted value. Updating the
functional property value using busctl set-property method is creating
a mismatch between persisted data and the data populated on DBus, as
set-property method updates the data only on the DBus. This is causing
the mismatch of the state of the FRU before and after the BMC reboot
while host is running, as DBus picks the value from the stale data
while host is running.

Updated to call Notify method of PIM service which updates data on both
persisted and on the DBus.

Signed-off-by: RekhaAparna01 <vrekhaaparna@ibm.com>